Initial commit.
Bootstrap of the website. The theme flatsum is the one selected, the landscape is the stock one that ships as example
I'm going to take it there for the moment to have some inspiration if needed. flatsum-html is a template I bought
because I like the timeline UI it provides.